extract SwordDocument object for creating the SWORD responses
